+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - Add :editable style to Table to enable auto-editing
 - `table` `header_visible` property to hide header when false
 - `table` `item_count` property to set minimum item count (fill empty rows when below in table items)
+- Support table selection data-binding
+- Fix issue with edit table item error on sorting table
 
 ## 0.6.1